payana/payana_core/requirements.txt
google-cloud-happybase==0.33.0
pyyaml
google-cloud-storage